fix:修复了一些逻辑错误

+++ app/user/domain/user.go
@@ -1,7 +1,7 @@
 package domain
 
 type User struct {
-	UserId   int32 
+	UserId   int64 
 	Username string 
 	Password string 
 }
